this is pretty much the end of the line

+++ src/programs/libprog/libc/string.h
@@ -2,7 +2,7 @@
 #include <stdbool.h>
 #include <stddef.h>
 
-void * memcpy(void * dest, const void * src, uint64_t len);
+void * memcpy(void * dest, const void * src, size_t len);
 char * strcpy(char * dest, const char * str);
 size_t strlen(const char * str);
 int strcmp(const char * str1, const char * str2);
